SEM15JUNIORTORRES: Restructuracion de Cliente
