+++ src/main/java/com/transport/management/modules/cliente/ClienteRepository.java
@@ -1,6 +1,5 @@
-package com.transport.management.repositories;
+package com.transport.management.modules.cliente;
 
-import com.transport.management.entities.ClienteEntity;
 import com.transport.management.utils.abtractBase.BaseRepository;
 
 public interface ClienteRepository extends BaseRepository<ClienteEntity> {
