+++ src/main/java/com/transport/management/services/ReniecService.java
@@ -8,7 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
-import com.transport.management.entities.ClienteEntity;
+import com.transport.management.modules.cliente.ClienteEntity;
 
 @Service
 public class ReniecService {
